+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/impl/RobotParameterImpl.java
@@ -11,7 +11,7 @@ import static dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTypes.*;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.*;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotPsiUtil;
 
-public class RobotParameterImpl extends RobotPsiElementBase implements RobotParameter {
+public class RobotParameterImpl extends RobotParameterExtension implements RobotParameter {
 
   public RobotParameterImpl(@NotNull ASTNode node) {
     super(node);