+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/annotator/compilation/RobotParameterAnnotator.java
@@ -11,11 +11,7 @@ import dev.xeonkryptos.xeonrobotframeworkplugin.config.RobotHighlighter;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedParameter;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotArgument;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotKeywordCall;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLocalSetting;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotParameter;
-import dev.xeonkryptos.xeonrobotframeworkplugin.util.KeyUtils;
-import dev.xeonkryptos.xeonrobotframeworkplugin.util.KeywordUtil;
-import dev.xeonkryptos.xeonrobotframeworkplugin.util.RobotNames;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
@@ -24,26 +20,21 @@ public class RobotParameterAnnotator extends RobotAnnotator {
 
     @Override
     public void visitParameter(@NotNull RobotParameter parameter) {
-        parameter.putUserData(KeyUtils.HANDLED_AS_SIMPLE_ARGUMENT_KEY, false);
-
-        RobotKeywordCall keywordCall = PsiTreeUtil.getParentOfType(parameter, RobotKeywordCall.class);
-        if (keywordCall != null) {
-            Collection<DefinedParameter> availableParameters = keywordCall.getAvailableParameters();
-            String parameterName = parameter.getParameterName();
-            boolean directMatchFound = availableParameters.stream().anyMatch(param -> param.matches(parameterName));
-            if (!directMatchFound && availableParameters.stream().noneMatch(DefinedParameter::isKeywordContainer)) {
-                keywordCall.getStartOfKeywordsOnlyIndex()
-                           .ifPresentOrElse(keywordsOnlyIndex -> handleParameterWithinKeywordOnlyKeywordCall(keywordsOnlyIndex, parameter, keywordCall, getHolder()), () -> {
-                               String keywordCallName = keywordCall.getName();
-                               String normalizedKeywordCallName = KeywordUtil.normalizeKeywordName(keywordCallName);
-                               // Create Dictionary keyword calls have special handling for parameters
-                               if (!RobotNames.CREATE_DICTIONARY_NORMALIZED_KEYWORD_NAME.equalsIgnoreCase(normalizedKeywordCallName)) {
-                                   convertParameterToArgumentVisually(getHolder(), parameter);
-                               }
-                           });
-            }
-        } else if (PsiTreeUtil.getParentOfType(parameter, RobotLocalSetting.class) != null) {
+        if (parameter.isFakeParameter()) {
             convertParameterToArgumentVisually(getHolder(), parameter);
+        } else {
+            RobotKeywordCall keywordCall = PsiTreeUtil.getParentOfType(parameter, RobotKeywordCall.class);
+            if (keywordCall != null) {
+                Collection<DefinedParameter> availableParameters = keywordCall.getAvailableParameters();
+                String parameterName = parameter.getParameterName();
+                if (keywordCall.getStartOfKeywordsOnlyIndex().isPresent()) {
+                    boolean noDirectMatchFound = availableParameters.stream().noneMatch(param -> param.matches(parameterName));
+                    boolean noKeywordContainerFound = availableParameters.stream().noneMatch(DefinedParameter::isKeywordContainer);
+                    if (noDirectMatchFound && noKeywordContainerFound) {
+                        keywordCall.getStartOfKeywordsOnlyIndex().ifPresent(keywordsOnlyIndex -> handleParameterWithinKeywordOnlyKeywordCall(keywordsOnlyIndex, parameter, keywordCall, getHolder()));
+                    }
+                }
+            }
         }
     }
 
@@ -79,6 +70,5 @@ public class RobotParameterAnnotator extends RobotAnnotator {
         // Grow range by 1 to include the assignment operator
         TextRange parameterIdTextRange = parameter.getParameterId().getTextRange().grown(1);
         holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES).range(parameterIdTextRange).textAttributes(RobotHighlighter.ARGUMENT).create();
-        parameter.putUserData(KeyUtils.HANDLED_AS_SIMPLE_ARGUMENT_KEY, true);
     }
 }