+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/util/KeyUtils.kt
@@ -1,9 +1,0 @@
-package dev.xeonkryptos.xeonrobotframeworkplugin.util
-
-import com.intellij.openapi.util.Key
-
-object KeyUtils {
-
-    @JvmField
-    val HANDLED_AS_SIMPLE_ARGUMENT_KEY = Key.create<Boolean>("HANDLED_AS_SIMPLE_ARGUMENT")
-}