+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Changed matching logic for code completion suggestions. They are based now on Camel Humps matching
 - Don't fold variables with a complex variable content (i.e. nested variables)
+- Removed support for special characters `*` and `/` at the start of a keyword. The insertion of mandatory parameters is handled by the quick fix already available. The insertion of optional
+  parameters is now available via a new intention
 
 ### Fixed